% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{argia_model_data}
\alias{argia_model_data}
\title{Argia Model Data}
\format{
An \code{sf} object with 550 rows and 15 columns:

\itemize{
\item SampleID: text
\item ProjectLabel: text
\item CollectionDate: text
\item presence: text
\item year: text
\item COMID: text
\item WSAREASWKM: text
\item ELEVWS: text
\item WETINDEXWS: text
\item BFIWS: text
\item PRECIP8110WS: text
\item PCTWETLANDXXXXWS: text
\item PCTIMPXXXXWSRP100: text
\item TMEANPRISMXXXXPT: text
\item geometry: \code{POINT} geometry representing coordinates in a
NAD83 projection (EPSG: 5070).
}
}
\usage{
argia_model_data
}
\description{
Argia Model Data
}
\keyword{datasets}
