% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{argia_pred_data_small}
\alias{argia_pred_data_small}
\title{Argia Prediction Data (Small)}
\format{
An \code{sf} object with 100 rows and 10 columns:

\itemize{
\item COMID: text
\item WSAREASWKM: text
\item ELEVWS: text
\item WETINDEXWS: text
\item BFIWS: text
\item PRECIP8110WS: text
\item PCTWETLANDXXXXWS: text
\item PCTIMPXXXXWSRP100: text
\item TMEANPRISMXXXXPT: text
\item geometry: \code{POINT} geometry representing coordinates in a
NAD83 projection (EPSG: 5070).
}
}
\usage{
argia_pred_data_small
}
\description{
Argia Prediction Data (Small)
}
\keyword{datasets}
