% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{comids}
\alias{comids}
\title{Argia Sample Location COMIDs}
\format{
A character vector of COMIDs for the sample locations in
\code{argia_model_data}.
}
\usage{
comids
}
\description{
Argia Sample Location COMIDs
}
\keyword{datasets}
