% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cond_model_data}
\alias{cond_model_data}
\title{Conductivity Model Data}
\format{
An \code{sf} object with 162 rows and 11 columns:

\itemize{
\item COMID: text
\item COND_RESULT: text
\item AREA_HA: text
\item year: text
\item TMEAN8110CAT: text
\item PRECIP8110WS: text
\item CAOWS: text
\item SWS: text
\item PCTCROPWS: text
\item PCTURBWS: text
\item geometry: \code{POINT} geometry representing coordinates in a
NAD83 projection (EPSG: 5070).
}
}
\usage{
cond_model_data
}
\description{
Conductivity Model Data
}
\keyword{datasets}
