% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cond_nla_data}
\alias{cond_nla_data}
\title{Conductivity NLA Data}
\format{
An \code{sf} object with 162 rows and 5 columns:

\itemize{
\item COMID: text
\item COND_RESULT: text
\item AREA_HA: text
\item year: text
\item geometry: \code{POINT} geometry representing coordinates in a
NAD83 projection (EPSG: 5070).
}
}
\usage{
cond_nla_data
}
\description{
Conductivity NLA Data
}
\keyword{datasets}
