% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spdata.sfs24package.R
\docType{package}
\name{data.sfs2024-package}
\alias{data.sfs2024}
\alias{data.sfs2024-package}
\title{data.sfs2024: Package that contains data for the Spatial Analysis and Statistical Modeling with R and spmodel workshop at the Society for Freshwater Science 2024 meeting}
\description{
What the package does (one paragraph).
}
\author{
\strong{Maintainer}: Michael Dumelle \email{Dumelle.Michael@epa.gov} (\href{https://orcid.org/0000-0002-3393-5529}{ORCID})

Other contributors:
\itemize{
  \item Ryan A. Hill \email{hill.ryan@epa.gov} (\href{https://orcid.org/0000-0001-9583-0426}{ORCID}) [contributor]
}

}
\keyword{internal}
