% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spdata.sfs24package.R
\docType{package}
\name{spdata.sfs24-package}
\alias{spdata.sfs24}
\alias{spdata.sfs24-package}
\title{spdata.sfs24: Package that contains data for the Spatial Analysis and Statistical Modeling with R and spmodel workshop at the Society for Freshwater Science 2024 meeting}
\description{
Contains data for the Spatial Analysis and Statistical Modeling with R and spmodel workshop at the Society for Freshwater Science 2024 meeting.
}
\author{
\strong{Maintainer}: Michael Dumelle \email{Dumelle.Michael@epa.gov} (\href{https://orcid.org/0000-0002-3393-5529}{ORCID})

Other contributors:
\itemize{
  \item Ryan A. Hill \email{hill.ryan@epa.gov} (\href{https://orcid.org/0000-0001-9583-0426}{ORCID}) [contributor]
}

}
\keyword{internal}
